refactor(import): extract monument feature mapping and enrichment to dedicated components
